+++ src/lib.rs
@@ -88,6 +88,11 @@ pub trait Filesystem {
 		Err(Error::from_raw_os_error(libc::ENOSYS))
 	}
 
+	fn rmdir(&mut self, _req: &Request, path: &Path) -> Result<()> {
+		let _ = path;
+		Err(Error::from_raw_os_error(libc::ENOSYS))
+	}
+
 	fn chown(&mut self, _req: &Request, path: &Path, uid: Option<u32>, gid: Option<u32>) -> Result<()> {
 		let _ = (path, uid, gid);
 		Err(Error::from_raw_os_error(libc::ENOSYS))