+++ ChangeLog.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Added
 
 - Filesystem::unlink()
+- Filesystem::rmdir()
 - Filesystem::chown()
 - Filesystem::chmod()
 - Filesystem::utime()